% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evapoTranspirationStats.R
\name{evapoTranspirationStats}
\alias{evapoTranspirationStats}
\title{Calculate evapo-transpiration stats from monthly weather data}
\usage{
evapoTranspirationStats(weather, siteInfo, sites)
}
\arguments{
\item{weather}{A data.frame with the following columns (names must match):
"site", "month", "year", "tmin", "tmax", "tmean", "rain".}

\item{siteInfo}{A dataframe with the column "site" that matches that of weather. This
dataset is populated with ET stats}

\item{sites}{A character vector specifying the names of sites in weathe / siteInfo}

\item{...}{Not currently in use.}
}
\value{
siteInfo, with three additional columns, spi, spei, pot.et
}
\description{
\code{evapoTranspirationStats} Simple function to clean up ET stats generation
from weather data.
}
\details{
See SPEI R package for details. This is just a simple wrapper.
}
\examples{
\dontrun{
# more here soon.
}

}
