% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genStatTools.R
\name{genstatEffect}
\alias{genstatEffect}
\alias{genstatLod}
\title{Parse genStat QTL model output}
\usage{
genstatEffect(path, pattern = "QTL-")
}
\arguments{
\item{path}{The path to the files to be parsed}

\item{pattern}{The pattern of filenames to look for}
}
\description{
\code{genstatEffect} calculates the effects from genstat models
}
\details{
parses QTL effects from genstat
}
