% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genStatTools.R
\name{genstatLod}
\alias{genstatLod}
\title{genstatLod}
\usage{
genstatLod(path, pattern = "LOD", cross)
}
\arguments{
\item{path}{The path to the files to be parsed}

\item{pattern}{The pattern of filenames to look for}

\item{cross}{The R/qtl cross object to be populated.}
}
\description{
\code{genstatLod} Extracts the -log10 P-value profile from genstat output
}
\details{
parses LOD scores from genstat
}
