% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genStatTools.R
\name{manualDrop}
\alias{manualDrop}
\title{manualDrop}
\usage{
manualDrop(s1, qtlMarkerList, drop = 1)
}
\arguments{
\item{s1}{The scanone-classed -log10 P-value curves (usually generated from genstatLod)}

\item{qtlMarkerList}{A list of marker names that represent the QTL of interest.}
}
\description{
\code{manualDrop} calculates CI from the -log10 P-value curves
}
\details{
manual calculation of confidence intervals
}
