% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\docType{data}
\name{station_data}
\alias{station_data}
\title{station_data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 587059 rows and 11 columns.}
\usage{
station_data
}
\references{
\url{http://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt}
}
\author{
JT Lovell \email{johntlovell@gmail.com}
}
\keyword{datasets}
